+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ### v0.0.13-SNAPSHOT - unreleased
  - introduce changelog
  - Remove `injector()` convenience method from main api
+ - introduce `mockspresso-extend` library
+ - refactor `mockspresso-quick` to build upon `mockspresso-extend`